requirements.txt
aiofiles==0.5.0
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.60.1
fs==2.4.11
h11==0.9.0
httptools==0.1.1
idna==2.10
Pillow==7.2.0
pkg-resources==0.0.0
pydantic==1.6.1
python-dotenv==0.14.0
python-multipart==0.0.5
pytz==2020.1
requests==2.24.0
six==1.15.0
starlette==0.13.6
urllib3==1.25.10
uvicorn==0.11.6
uvloop==0.14.0
websockets==8.1
